requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.8
aiohttp==3.11.13
aiosignal==1.3.2
altgraph==0.17.4
appdirs==1.4.4
asyncgui==0.6.3
asyncio==3.4.3
asynckivy==0.6.4
attrs==25.1.0
babel==2.16.0
beautifulsoup4==4.13.4
blinker==1.9.0
bottle==0.13.2
bs4==0.0.2
build==1.3.0
buildozer @ git+https://github.com/kivy/buildozer.git@1e3d533f5ebb7b5d86392c936c8de852d7110eeb
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cryptography==44.0.2
cssselect==1.3.0
cycler==0.12.1
Cython==0.29.36
dailymotion==0.2.5
dash==2.18.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.2.1
distlib==0.4.0
dnspython==2.7.0
docutils==0.21.2
email_validator==2.2.0
et_xmlfile==2.0.0
fake-useragent==2.2.0
ffmpeg-python==0.2.0
filelock==3.17.0
filetype==1.2.0
Flask==3.0.3
fonttools==4.55.4
frozenlist==1.5.0
fsspec==2024.12.0
future==1.0.0
greenlet==3.2.4
hf-xet==1.1.5
huggingface-hub==0.33.4
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
imbalanced-learn==0.14.0
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
Kivy==2.3.1
Kivy-Garden==0.1.5
kivymd @ https://github.com/kivymd/KivyMD/archive/master.zip#sha256=ab18d4cc36e2d53ea87f463d01299c39c75bae66cc1950f9bc6397f23317c848
kivymd-extensions==1.0.0
kiwisolver==1.4.8
lightgbm==4.6.0
lxml==6.0.1
lxml_html_clean==0.4.2
MarkupSafe==3.0.2
materialyoucolor==2.0.10
matplotlib==3.10.0
meson==1.9.1
MouseInfo==0.1.3
moviepy==2.2.1
mpmath==1.3.0
multidict==6.1.0
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.13.0
nltk==3.8.1
Nuitka==2.8.4
numpy==2.2.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-python==4.12.0.88
openpyxl==3.1.5
ordered-set==4.1.0
packaging==24.2
pandas==2.2.3
parse==1.20.2
paypalrestsdk==1.13.3
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.5.0
playwright==1.54.0
plotly==5.24.1
plyer==2.1.0
proglog==0.1.12
propcache==0.3.0
proxy_tools==0.1.0
ptyprocess==0.7.0
PyAutoGUI==0.9.54
pycparser==2.22
pyee==13.0.0
pygame==2.6.1
PyGetWindow==0.0.9
Pygments==2.19.1
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2025.1
pyjnius==1.6.1
PyMsgBox==2.0.1
PyMuPDF==1.25.3
pyOpenSSL==25.0.0
pyoxidizer==0.24.0
pyparsing==3.2.1
pyperclip==1.9.0
pyppeteer==2.0.0
pyproject_hooks==1.2.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.16
PyQt5_sip==12.17.0
pyquery==2.0.1
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-for-android @ git+https://github.com/kivy/python-for-android.git@957a3e5f8c270f7aa648ba185e5a68c1077a798d
python-mpv==1.0.8
python-vlc==3.0.21203
python3-xlib==0.15
pytweening==1.2.0
pytz==2024.2
PyVirtualDisplay==3.0
pywebview==5.4
PyYAML==6.0.2
QtPy==2.4.3
regex==2024.11.6
reportlab==4.4.4
requests==2.32.3
requests-html==0.10.0
requests-toolbelt==1.0.0
retrying==1.3.4
safetensors==0.5.3
schedule==1.2.2
scikit-learn==1.6.1
scipy==1.15.2
SCons==4.8.1
seaborn==0.13.2
setuptools==80.9.0
sh==1.14.3
six==1.17.0
sounddevice==0.5.2
soupsieve==2.7
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
tkcalendar==1.6.1
tokenizers==0.21.2
toml==0.10.2
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tqdm==4.67.1
transformers==4.53.2
triton==3.1.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==1.26.20
virtualenv==20.35.4
w3lib==2.3.1
websockets==10.4
webview-android==0.10
Werkzeug==3.0.6
wheel==0.43.0
yarl==1.18.3
yt-dlp==2025.1.26
zipp==3.21.0
zstandard==0.23.0
